+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1015](https://github.com/spegel-org/spegel/pull/1015) Remove explicit GOMAXPROCS
 - [#1017](https://github.com/spegel-org/spegel/pull/1017) Use prefer same node traffic distribution.
 - [#1018](https://github.com/spegel-org/spegel/pull/1018) Add optional default values when parting image reference.
+- [#1022](https://github.com/spegel-org/spegel/pull/1022) Refactor and rename fingerprint media type.
 
 ### Deprecated